requirements.txt
Flask
gunicorn
PyPDF2
python-docx
spacy